fixed timestamp and CMakelists

+++ src/ichthus_lidar_driver/common/lidar.cpp
@@ -306,8 +306,8 @@ namespace VLP16
                 // point.ts = msg.ts.tv_sec();
                 if (need_update == true)
                 {
-                  sec = (float)((int)(*vector_msg[i]).ts.sec % 1000000); // fixed by qjin, 1. long->float , 2. upper value not needed
-                  usec = (float)((int)(*vector_msg[i]).ts.nsec / 1000);  // fixed by qjin, 1. long->float
+                  sec = (float)((int)((*vector_msg[i]).ts.sec) % 1000000); // fixed by qjin, 1. long->float , 2. upper value not needed
+                  usec = (float)((int)((*vector_msg[i]).ts.nsec) / 1000);  // fixed by qjin, 1. long->float
                   need_update = false;
                 }
                 point.normal_y = sec;
@@ -451,8 +451,8 @@ namespace OSI64
 
           if (need_update == true)
           {
-            sec = (float)(msg.ts.sec % 1000000); // fixed by qjin, 1. long->float , 2. upper value not needed
-            usec = (float)(msg.ts.nsec / 1000); // fixed by qjin, 1. long->float
+            sec = (float)((int)(msg.ts.sec) % 1000000); // fixed by qjin, 1. long->float , 2. upper value not needed
+            usec = (float)((int)(msg.ts.nsec) / 1000); // fixed by qjin, 1. long->float
             need_update = false;
           }
           point.normal_y = sec;
